Refactor plotting codes
