+++ src/cartogram_info/cartogram_info.cpp
@@ -437,7 +437,7 @@ void CartogramInfo::write_svg(const std::string &suffix)
   for (const InsetState &inset_state : inset_states_) {
     inset_names += inset_state.pos();
   }
-  insets_combined.write_cairo_map(
+  insets_combined.write_map(
     map_name_ + "_" + inset_names + "_" + suffix,
     false);
 }
