+++ src/inset_state/check_topology.cpp
@@ -39,7 +39,7 @@ void InsetState::is_simple(const char *caller_func) const
         std::cerr << "ERROR: Outer boundary is not simple for GeoDiv "
                   << gd.id();
         std::cerr << ". is_simple() called from " << caller_func << std::endl;
-        write_cairo_map(
+        write_map(
           inset_name_ + "_" + std::to_string(n_finished_integrations_) +
             "_not_simple_after_" + caller_func,
           false);
@@ -50,7 +50,7 @@ void InsetState::is_simple(const char *caller_func) const
           std::cerr << "ERROR: Hole is not simple for GeoDiv " << gd.id();
           std::cerr << ". is_simple() called from " << caller_func
                     << std::endl;
-          write_cairo_map(
+          write_map(
             inset_name_ + "_" + std::to_string(n_finished_integrations_) +
               "_not_simple_after_" + caller_func,
             false);
