+++ src/inset_state/fill_with_density.cpp
@@ -16,6 +16,6 @@ void InsetState::fill_with_density() {
   // Plot density map if requested
   if (args_.plot_density) {
       std::string file_name = file_prefix_ + "_unblurred_density.svg";
-      write_density_image(file_name, false);
+      write_density_image(file_name);
   }
 }
