+++ src/inset_state/integrate.cpp
@@ -33,7 +33,7 @@ void InsetState::preprocess()
     // Color if necessary
     auto_color();
     initialize_identity_proj();
-    write_cairo_map(inset_name_ + "_input", args_.plot_grid, true);
+    write_map(inset_name_ + "_input", args_.plot_grid, true);
   }
 
   timer.stop("Preprocessing");
@@ -135,7 +135,7 @@ void InsetState::integrate(ProgressTracker &progress_tracker)
 
   // Write SVG for this inset, if requested
   if (args_.plot_polygons) {
-    write_cairo_map(inset_name() + "_output", args_.plot_grid, false);
+    write_map(inset_name() + "_output", args_.plot_grid, false);
   }
 
   // Project original map with cumulative projection
